update RequestParser class variable name to rawRequest and remove redundant RouteTemplate class
