+++ src/main/java/server/router/routes/NotFound.java
@@ -1,37 +1,22 @@
 package server.router.routes;
 
-import java.util.List;
 import server.constants.Status;
 import server.request.Request;
 import server.response.Response;
 import server.response.ResponseBuilder;
 
-public class NotFound extends Route {
+public class NotFound {
+  private final Request request;
+
   public NotFound(Request request) {
-    super(request);
+    this.request = request;
   }
 
-  @Override
-  public String path() {
-    return null;
-  }
-
-  @Override
-  public List<String> methods() {
-    return null;
-  }
-
-  @Override
-  public String resource() {
-    return null;
-  }
-
-  @Override
   public Response processRequest() {
     StringBuilder startLine = new StringBuilder();
     startLine.append(request.protocol());
     startLine.append(" ");
-    startLine.append(Status.NOT_FOUND.toString());
+    startLine.append(Status.NOT_FOUND);
 
     return new ResponseBuilder()
                 .setStartLine(startLine.toString())
